Rename sequences
